Make data storage headers private
