+++ packages/arb-avm-cpp/data_storage/include/data_storage/checkpointstorage.hpp
@@ -26,6 +26,8 @@
 
 struct GetResults;
 class Machine;
+class BlockStore;
+class ConfirmedNodeStore;
 
 namespace rocksdb {
 class TransactionDB;
